+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * key "auto" for each-block
 * anchors for components
 * style templates, e.g. style:color={color}, style:border-color="red", style:border="1px solid {color}"
+* compile option "css"
 
 ## 0.5.x